+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of threedigrid-builder
 1.7.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Replaced threedi-modelchecker dependency with threedi-schema==214.*.
 
 
 1.7.0 (2023-01-11)